rename header components to navigation
